# proyecto-gimnasio/backend/docs/swagger.yaml
basePath: /
definitions:
  models.SwaggerActivity:
    properties:
      capacity:
        type: integer
      category_id:
        type: integer
      day:
        type: string
      duration:
        type: integer
      frequency:
        type: string
      id:
        type: integer
      image_url:
        type: string
      name:
        type: string
      start_hour:
        type: string
    type: object
  models.SwaggerCategory:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  models.SwaggerUser:
    properties:
      id:
        type: integer
      password:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: API Support
    url: http://www.swagger.io/support
  description: API para la gestión de un gimnasio.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Proyecto Gimnasio API
  version: "1.0"
paths:
  /activities:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.SwaggerActivity'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Listar todas las actividades
      tags:
      - Actividades
  /activities/{id}:
    get:
      parameters:
      - description: ID de la actividad
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SwaggerActivity'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Obtener una actividad por ID
      tags:
      - Actividades
  /activities/enroll:
    post:
      consumes:
      - application/json
      parameters:
      - description: user_id y activity_id
        in: body
        name: inscripcion
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Inscribir usuario en una actividad
      tags:
      - Inscripciones
  /admin/activities:
    post:
      consumes:
      - application/json
      parameters:
      - description: Datos de la actividad
        in: body
        name: activity
        required: true
        schema:
          $ref: '#/definitions/models.SwaggerActivity'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Crear una nueva actividad
      tags:
      - Actividades
  /admin/activities/{id}:
    delete:
      parameters:
      - description: ID de la actividad
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Eliminar una actividad
      tags:
      - Actividades
    put:
      consumes:
      - application/json
      parameters:
      - description: ID de la actividad
        in: path
        name: id
        required: true
        type: integer
      - description: Datos actualizados
        in: body
        name: activity
        required: true
        schema:
          $ref: '#/definitions/models.SwaggerActivity'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Actualizar una actividad existente
      tags:
      - Actividades
  /categories:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.SwaggerCategory'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Listar todas las categorías
      tags:
      - Categorías
    post:
      consumes:
      - application/json
      parameters:
      - description: Datos de la categoría
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/models.SwaggerCategory'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.SwaggerCategory'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Crear una nueva categoría
      tags:
      - Categorías
  /categories/{id}:
    delete:
      parameters:
      - description: ID de la categoría
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Eliminar una categoría
      tags:
      - Categorías
    put:
      consumes:
      - application/json
      parameters:
      - description: ID de la categoría
        in: path
        name: id
        required: true
        type: integer
      - description: Datos actualizados
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/models.SwaggerCategory'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SwaggerCategory'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Actualizar una categoría
      tags:
      - Categorías
  /login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Credenciales de usuario
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/models.SwaggerUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Iniciar sesión
      tags:
      - Autenticación
  /register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Datos del usuario
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.SwaggerUser'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Registrar nuevo usuario
      tags:
      - Autenticación
  /users/{user_id}/activities:
    get:
      parameters:
      - description: ID del usuario
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.SwaggerActivity'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Ver actividades de un usuario
      tags:
      - Inscripciones
swagger: "2.0"